+++ src/plugins/clean-html/helpers/visitor/filters/index.ts
@@ -13,9 +13,13 @@
  */
 
 export * from './allow-attributes';
+export * from './convert-unsafe-embeds';
 export * from './fill-empty-paragraph';
 export * from './remove-empty-text-node';
 export * from './remove-inv-text-nodes';
 export * from './replace-old-tags';
+export * from './safe-links-target';
+export * from './sandbox-iframes-in-content';
 export * from './sanitize-attributes';
+export * from './sanitize-styles';
 export * from './try-remove-node';