+++ src/types/ui.d.ts
@@ -26,7 +26,7 @@ export interface IUIElement
 	parentElement: Nullable<IUIElement>;
 	container: HTMLElement;
 	name: string;
-	
+
 	getRole(): string;
 
 	/**